pyproject.toml
[project]
name = "video-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "langchain>=1.0.5",
    "langchain-ollama>=1.0.0",
    "openai-whisper>=20250625",
]
